Open a file or directory

Functions to open a file with its default program or open a file's location in the file explorer.

Usage

open_file(path)

open_location(path)

file.open(path)

dir.open(path)

Arguments

path

The path to the file or directory.

Details

file.open() and dir.open() are aliases, in line with base::file.create, file.exists, dir.create, dir.exists, etc.

Functions

open_file(): Opens the file at the specified path using the default program for that file type. Automatically handles paths with special characters.

open_location(): Opens the file explorer at the location of the specified file, with the file selected if possible. Automatically handles paths with special characters.

Examples

if (FALSE) { # \dontrun{
# Open a file
open_file("path/to/file.txt")

# Open a file's location
open_location("path/to/file.txt")
} # }