Custom summary table

Generates a summary table, with a row for each variable passed to .vars and a column for each function passed to ....

Usage

summary_table(
  .data,
  ...,
  .vars = where(is.numeric),
  na.rm = FALSE,
  .rows_group_by = NULL,
  .cols_group_by = NULL,
  .cols_group_opts = list(),
  .var_col_name = "Variable"
)

Arguments

.data

A data frame or data frame extension (e.g. a tibble).

...

Functions to apply to each variable specified in .vars, as function names (e.g., mean) or purrr-style lambdas (e.g., ~ mean(.x, na.rm = TRUE)). Each function passed to ... will yield one column in the output table (or one column per group if .cols_group_by is specified). Output column names can optionally be specified. E.g., m = mean, sd, sem = ~ sd(.x) / sqrt(n()) yields output columns named m, sd, and sem.

.vars

<tidy-select> Columns in .data to summarize. Each column passed to .vars will yield one row in the output table (or one row per group if .rows_group_by is specified).

na.rm

A logical value passed to any functions in ... that take an na.rm argument.

.rows_group_by

<tidy-select> Grouping variable(s) for output rows.

.cols_group_by

<tidy-select> Grouping variable(s) for output columns.

.cols_group_opts

A list of arguments passed to lighthouse::pivot_wider_alt() if .cols_group_by is specified.

.var_col_name

The name of the output column containing variable names passed to .vars.

Examples

# example data
mtcars2 <- mtcars %>%
  dplyr::mutate(Transmission = recode(am, `0` = "auto", `1` = "manual"))
#> Error in mtcars %>% dplyr::mutate(Transmission = recode(am, `0` = "auto",     `1` = "manual")): could not find function "%>%"

# simple summary table. note specification of column and row names
# for "n", "m", and "weight".
mtcars2 %>%
  summary_table(
    n = ~ sum(!is.na(.x)), m = mean, sd,
    .vars = c(mpg, hp, weight = wt)
  )
#> Error in mtcars2 %>% summary_table(n = ~sum(!is.na(.x)), m = mean, sd,     .vars = c(mpg, hp, weight = wt)): could not find function "%>%"

# with column and row groupings
mtcars2 %>%
  summary_table(
    n = ~ sum(!is.na(.x)), m = mean, sd,
    .vars = c(mpg, hp, weight = wt),
    .cols_group_by = cyl,
    .rows_group_by = Transmission
  )
#> Error in mtcars2 %>% summary_table(n = ~sum(!is.na(.x)), m = mean, sd,     .vars = c(mpg, hp, weight = wt), .cols_group_by = cyl, .rows_group_by = Transmission): could not find function "%>%"

# customize output column names with `.cols_group_opts`
mtcars2 %>%
  summary_table(
    n = ~ sum(!is.na(.x)), m = mean, sd,
    .vars = c(mpg, hp, weight = wt),
    .cols_group_by = cyl,
    .rows_group_by = Transmission,
    .cols_group_opts = list(
      names_value_first = FALSE,
      names_prefix = "cyl",
      names_value_sep = ": "
    )
  )
#> Error in mtcars2 %>% summary_table(n = ~sum(!is.na(.x)), m = mean, sd,     .vars = c(mpg, hp, weight = wt), .cols_group_by = cyl, .rows_group_by = Transmission,     .cols_group_opts = list(names_value_first = FALSE, names_prefix = "cyl",         names_value_sep = ": ")): could not find function "%>%"

# can alternatively pass a glue specification to `.cols_group_opts`
mtcars2 %>%
  summary_table(
    n = ~ sum(!is.na(.x)), m = mean, sd,
    .vars = c(mpg, hp, weight = wt),
    .cols_group_by = cyl,
    .rows_group_by = Transmission,
    .cols_group_opts = list(names_glue = "{.value}: {cyl}-cyl")
  )
#> Error in mtcars2 %>% summary_table(n = ~sum(!is.na(.x)), m = mean, sd,     .vars = c(mpg, hp, weight = wt), .cols_group_by = cyl, .rows_group_by = Transmission,     .cols_group_opts = list(names_glue = "{.value}: {cyl}-cyl")): could not find function "%>%"