Determine fiscal year from date

Given a date, returns the corresponding fiscal year, start date, or end date. The fiscal_year function allows specifying the fiscal year start month, while ffy and sfy_il are convenience wrappers:

ffy: Federal fiscal year (starts in October)

sfy_il: Illinois state fiscal year (starts in July)

Usage

fiscal_year(x, type = c("year", "date_first", "date_last"), fiscal_start = 1)

ffy(x, type = c("year", "date_first", "date_last"))

sfy_il(x, type = c("year", "date_first", "date_last"))

Arguments

x

A date or date-time object.

type

What to return: the fiscal year ("year"), first day of the fiscal year ("date_first"), or last day of the fiscal year ("date_last").

fiscal_start

For fiscal_year, the month the fiscal year starts (default is 1 for January).

Value

An integer representing the fiscal year or a Date representing the start or end of the fiscal year, depending on type.

Examples

dt <- as.Date("2023-08-15")
fiscal_year(dt)
#> [1] 2023
fiscal_year(dt, fiscal_start = 7)
#> [1] 2024
fiscal_year(dt, type = "date_first", fiscal_start = 7)
#> [1] "2023-07-01"
fiscal_year(dt, type = "date_last", fiscal_start = 7)
#> [1] "2024-06-30"
ffy(dt)
#> [1] 2023
sfy_il(dt)
#> [1] 2024