Alternative column ordering and naming for pivot_wider()

A wrapper around tidyr::pivot_wider() with additional options for sorting and naming of output columns, with arguments sort_by_col, names_value_first, and names_value_sep These options are only relevant when more than one input column is passed to values_from.

Usage

pivot_wider_alt(
  data,
  id_cols = NULL,
  names_from = name,
  sort_by_col = TRUE,
  names_value_first = TRUE,
  names_value_sep = ".",
  names_sep = "_",
  names_prefix = "",
  names_glue = NULL,
  names_repair = "check_unique",
  values_from = value,
  values_fill = NULL,
  values_fn = NULL
)

Arguments

data

A data frame to pivot.

id_cols

<tidy-select> A set of columns that uniquely identifies each observation. Defaults to all columns in data except for the columns specified in names_from and values_from. Typically used when you have redundant variables, i.e. variables whose values are perfectly correlated with existing variables.

names_from

<tidy-select> A pair of arguments describing which column (or columns) to get the name of the output column (names_from), and which column (or columns) to get the cell values from (values_from).

If values_from contains multiple values, the value will be added to the front of the output column.

sort_by_col

If TRUE (default), output columns will be sorted by names_from, then values_from. (Differs from tidyr::pivot_wider(), which sorts by values_from first, then names_from.)

names_value_first

If FALSE, output columns will be named using a {column}_{.value} scheme. (Differs from tidyr::pivot_wider(), which uses a {.value}_{column} scheme.)

names_value_sep, names_sep

If names_from or values_from contain multiple variables, these will be used to join their values together into a single string to use as a column name. names_value_sep will separate {.value} from {column} components, while names_sep will separate {column} components from one another if names_from contains multiple variables. See Details and Examples.

names_prefix, names_glue, names_repair, values_from, values_fill, values_fn

See documentation at tidyr::pivot_wider().

Details

The names_value_sep argument allows output column names that use a different separator between {.value} and {column} than between multiple {columns}s. Example:

pivot_wider_alt(
  fakedata,
  names_from = c(size, color), # size = "sm", "med", "lg"; color = "red", "blue"
  values_from = c(n, weight),
  names_sep = "_",
  names_value_sep = ": "
)

# output column names:
# `n: sm_red`, `weight: sm_red`, `n: sm_blue`, `weight: sm_blue`, `n: med_red`...

Examples

data_ex <- ggplot2::diamonds %>%
  group_by(cut, color) %>%
  summarize(Min = min(price), Median = median(price), Max = max(price))
#> `summarise()` has grouped output by 'cut'. You can override using the `.groups`
#> argument.

# default pivot_wider() behavior
data_ex %>%
  pivot_wider(
    id_cols = color,
    names_from = cut,
    values_from = Min:Max
  )
#> Error in pivot_wider(., id_cols = color, names_from = cut, values_from = Min:Max): could not find function "pivot_wider"

# pivot_wider_alt() behavior
data_ex %>%
  pivot_wider_alt(
    id_cols = color,
    names_from = cut,
    values_from = Min:Max
)
#> # A tibble: 7 × 16
#>   color Min.Fair Median.Fair Max.Fair Min.Good Median.Good Max.Good
#>   <ord>    <int>       <dbl>    <int>    <int>       <dbl>    <int>
#> 1 D          536        3730    16386      361       2728.    18468
#> 2 E          337        2956    15584      327       2420     18236
#> 3 F          496        3035    17995      357       2647     18686
#> 4 G          369        3057    18574      394       3340     18788
#> 5 H          659        3816    18565      368       3468.    18640
#> 6 I          735        3246    18242      351       3640.    18707
#> 7 J          416        3302    18531      335       3733     18325
#> # … with 9 more variables: `Min.Very Good` <int>, `Median.Very Good` <dbl>,
#> #   `Max.Very Good` <int>, Min.Premium <int>, Median.Premium <dbl>,
#> #   Max.Premium <int>, Min.Ideal <int>, Median.Ideal <dbl>, Max.Ideal <int>

# with `names_value_first` = FALSE
data_ex %>%
  pivot_wider_alt(
    id_cols = color,
    names_from = cut,
    values_from = Min:Max,
    names_value_first = FALSE
  )
#> # A tibble: 7 × 16
#>   color Fair.Min Fair.Median Fair.Max Good.Min Good.Median Good.Max
#>   <ord>    <int>       <dbl>    <int>    <int>       <dbl>    <int>
#> 1 D          536        3730    16386      361       2728.    18468
#> 2 E          337        2956    15584      327       2420     18236
#> 3 F          496        3035    17995      357       2647     18686
#> 4 G          369        3057    18574      394       3340     18788
#> 5 H          659        3816    18565      368       3468.    18640
#> 6 I          735        3246    18242      351       3640.    18707
#> 7 J          416        3302    18531      335       3733     18325
#> # … with 9 more variables: `Very Good.Min` <int>, `Very Good.Median` <dbl>,
#> #   `Very Good.Max` <int>, Premium.Min <int>, Premium.Median <dbl>,
#> #   Premium.Max <int>, Ideal.Min <int>, Ideal.Median <dbl>, Ideal.Max <int>

# multiple `names_from` vars, with different value vs. name separators
ggplot2::mpg %>%
  filter(class %in% c("compact", "subcompact", "midsize")) %>%
  group_by(manufacturer, trans = str_extract(trans, ".*(?=\\()"), year) %>%
  summarize(across(c(cty, hwy), mean)) %>%
  pivot_wider_alt(
    names_from = trans:year,
    values_from = cty:hwy,
    names_sep = "_",
    names_value_sep = ": "
  )
#> Error in group_by(., manufacturer, trans = str_extract(trans, ".*(?=\\()"),     year): Problem adding computed columns.
#> Caused by error in `mutate()`:
#> ! Problem while computing `trans = str_extract(trans, ".*(?=\\()")`.
#> Caused by error in `str_extract()`:
#> ! could not find function "str_extract"

# cleanup
rm(data_ex)